+++ baseband/vlbi_base/frame.py
@@ -1,1 +1,0 @@
-from ..base.frame import FrameBase as VLBIFrameBase  # noqa