+++ baseband/vlbi_base/header.py
@@ -1,4 +1,0 @@
-from ..base.header import (  # noqa
-    fixedvalue, four_word_struct, eight_word_struct,
-    make_parser, make_setter, get_default,
-    ParserDict, HeaderParser, VLBIHeaderBase)