+++ baseband/vlbi_base/payload.py
@@ -1,1 +1,0 @@
-from ..base.payload import PayloadBase as VLBIPayloadBase  # noqa